Fixed the issue:
Showed seemingly random behaviour, this was due to gen_cell() failing to detect
when a position was already occupied, due to the 'continue' only skipping within
the 'for loop' where as it was intended for skipping in the 'while loop'. Thus by
using a variable we are able to overcome this issue, and all seems to be behaving
correctly now.